requirements.txt
customtkinter==5.2.2
numpy==2.2.4
Pillow==11.2.0
CustomTkinterMessagebox==0.0.4
